+++ src/main/java/pl/coderslab/cart_item/CartItemRepository.java
@@ -5,6 +5,8 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import pl.coderslab.cart.Cart;
 
+import java.util.List;
+
 @Repository
 public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 
@@ -14,4 +16,6 @@ public interface CartItemRepository extends JpaRepository<CartItem, Long> {
     @Query ("select SUM(ci.product.price * ci.count) from CartItem ci join Product p on p.id = ci.product.id\n" +
             "                  where ci.cart.id = ?1")
     Double totalPriceCartItemsByUserId(Long carId);
+
+    List<CartItem> findCartItemByCartId(Long id);
 }
