+++ src/main/java/pl/coderslab/cart_item/JpaCartItemService.java
@@ -4,6 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 @Repository
 @RequiredArgsConstructor
@@ -27,5 +29,15 @@ public class JpaCartItemService implements CartItemService{
     @Override
     public void editCartItem(CartItem cartItem) {
         cartItemRepository.save(cartItem);
+    }
+
+    public CartItem getCartItemByProductIdAndCartId(Long productId, Long userId){
+        return cartItemRepository.findByProductIdAndCartId(productId, userId);
+    }
+    public Double getTotalPriceCartItemsByUserId(Long userId){
+        return cartItemRepository.totalPriceCartItemsByUserId(userId);
+    }
+    public List<CartItem> getCartItemByCartId(Long cartId){
+        return cartItemRepository.findCartItemByCartId(cartId);
     }
 }
