+++ src/main/webapp/js/counter.js
@@ -1,6 +1,7 @@
 
 document.addEventListener("DOMContentLoaded",function(){
 
+
 window.addEventListener('click', function(event){
   if(event.target.dataset.action === 'plus'){
     const alertTotal = event.target.closest('.alert');
@@ -11,6 +12,10 @@ window.addEventListener('click', function(event){
     const totalPrice = alertTotal.querySelector('.total-price');
     const id = alertTotal.querySelector('.id');
     totalPrice.innerText = "$" + ((parseFloat((counter.innerText).replace('$', ''))) * (parseFloat((priceOfProduct.innerText).replace('$', '')))).toString();
+      const dataTogle2 = this.document.querySelector('#dropdown');
+      const dataTogle = dataTogle2.children[1].children[0].innerText;
+      const allCount = this.document.querySelectorAll('[data-counter]');
+      dataTogle.innerText = countAll(allCount);
     addInformation(id, counter);
   
 }
@@ -42,4 +47,16 @@ function addInformation(id, counter){
 
 function deleteRow(id){
   fetch('/delete/' + id.value).then(function(){console.log("error")});
+}
+function countAll(allCount){
+    console.log("allcount",allCount);
+    const newArr = [];
+  const arr = Array.from(allCount).forEach(element => {
+  newArr.push(parseInt(element.innerText));
+  });
+  let sum = 0;
+  for (let i=0; i<newArr.length; i++) {
+    sum += newArr[i];
+  }
+  return sum;
 }
